+++ docs/search/variables_9.js
@@ -1,4 +1,5 @@
 var searchData=
 [
-  ['secret_0',['secret',['../structsensory_1_1token__manager_1_1_access_token_credentials.html#ad45fdbcbe828045fde84f82433ddbcd6',1,'sensory::token_manager::AccessTokenCredentials']]]
+  ['secret_0',['secret',['../structsensory_1_1token__manager_1_1_access_token_credentials.html#ad45fdbcbe828045fde84f82433ddbcd6',1,'sensory::token_manager::AccessTokenCredentials']]],
+  ['sections_1',['sections',['../classsensory_1_1io_1_1_i_n_i_reader.html#a4dd99b00c7c60eabc9a455a655c13a2e',1,'sensory::io::INIReader']]]
 ];